requirements.txt
click==7.1.2
Flask==1.1.2
flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
gunicorn==20.0.4
Pillow==2.2.2
requests==2.23.0
cloudinary==1.21.1